Create site selection controller
