+++ app/views/candidate_interface/course_choices/replace_choices/site_selection/replace_location.html.erb
